+++ lib/src/service/tweets/tweet_meta.freezed.dart
@@ -28,6 +28,8 @@ mixin _$TweetMeta {
   int? get resultCount => throw _privateConstructorUsedError;
   @JsonKey(name: 'next_token')
   String? get nextToken => throw _privateConstructorUsedError;
+  @JsonKey(name: 'previous_token')
+  String? get previousToken => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
@@ -43,7 +45,8 @@ abstract class $TweetMetaCopyWith<$Res> {
       {@JsonKey(name: 'newest_id') String? newestId,
       @JsonKey(name: 'oldest_id') String? oldestId,
       @JsonKey(name: 'result_count') int? resultCount,
-      @JsonKey(name: 'next_token') String? nextToken});
+      @JsonKey(name: 'next_token') String? nextToken,
+      @JsonKey(name: 'previous_token') String? previousToken});
 }
 
 /// @nodoc
@@ -60,6 +63,7 @@ class _$TweetMetaCopyWithImpl<$Res> implements $TweetMetaCopyWith<$Res> {
     Object? oldestId = freezed,
     Object? resultCount = freezed,
     Object? nextToken = freezed,
+    Object? previousToken = freezed,
   }) {
     return _then(_value.copyWith(
       newestId: newestId == freezed
@@ -78,6 +82,10 @@ class _$TweetMetaCopyWithImpl<$Res> implements $TweetMetaCopyWith<$Res> {
           ? _value.nextToken
           : nextToken // ignore: cast_nullable_to_non_nullable
               as String?,
+      previousToken: previousToken == freezed
+          ? _value.previousToken
+          : previousToken // ignore: cast_nullable_to_non_nullable
+              as String?,
     ));
   }
 }
@@ -92,7 +100,8 @@ abstract class _$TweetMetaCopyWith<$Res> implements $TweetMetaCopyWith<$Res> {
       {@JsonKey(name: 'newest_id') String? newestId,
       @JsonKey(name: 'oldest_id') String? oldestId,
       @JsonKey(name: 'result_count') int? resultCount,
-      @JsonKey(name: 'next_token') String? nextToken});
+      @JsonKey(name: 'next_token') String? nextToken,
+      @JsonKey(name: 'previous_token') String? previousToken});
 }
 
 /// @nodoc
@@ -110,6 +119,7 @@ class __$TweetMetaCopyWithImpl<$Res> extends _$TweetMetaCopyWithImpl<$Res>
     Object? oldestId = freezed,
     Object? resultCount = freezed,
     Object? nextToken = freezed,
+    Object? previousToken = freezed,
   }) {
     return _then(_TweetMeta(
       newestId: newestId == freezed
@@ -128,6 +138,10 @@ class __$TweetMetaCopyWithImpl<$Res> extends _$TweetMetaCopyWithImpl<$Res>
           ? _value.nextToken
           : nextToken // ignore: cast_nullable_to_non_nullable
               as String?,
+      previousToken: previousToken == freezed
+          ? _value.previousToken
+          : previousToken // ignore: cast_nullable_to_non_nullable
+              as String?,
     ));
   }
 }
@@ -139,7 +153,8 @@ class _$_TweetMeta implements _TweetMeta {
       {@JsonKey(name: 'newest_id') this.newestId,
       @JsonKey(name: 'oldest_id') this.oldestId,
       @JsonKey(name: 'result_count') this.resultCount,
-      @JsonKey(name: 'next_token') this.nextToken});
+      @JsonKey(name: 'next_token') this.nextToken,
+      @JsonKey(name: 'previous_token') this.previousToken});
 
   factory _$_TweetMeta.fromJson(Map<String, dynamic> json) =>
       _$$_TweetMetaFromJson(json);
@@ -156,10 +171,13 @@ class _$_TweetMeta implements _TweetMeta {
   @override
   @JsonKey(name: 'next_token')
   final String? nextToken;
+  @override
+  @JsonKey(name: 'previous_token')
+  final String? previousToken;
 
   @override
   String toString() {
-    return 'TweetMeta(newestId: $newestId, oldestId: $oldestId, resultCount: $resultCount, nextToken: $nextToken)';
+    return 'TweetMeta(newestId: $newestId, oldestId: $oldestId, resultCount: $resultCount, nextToken: $nextToken, previousToken: $previousToken)';
   }
 
   @override
@@ -171,7 +189,9 @@ class _$_TweetMeta implements _TweetMeta {
             const DeepCollectionEquality().equals(other.oldestId, oldestId) &&
             const DeepCollectionEquality()
                 .equals(other.resultCount, resultCount) &&
-            const DeepCollectionEquality().equals(other.nextToken, nextToken));
+            const DeepCollectionEquality().equals(other.nextToken, nextToken) &&
+            const DeepCollectionEquality()
+                .equals(other.previousToken, previousToken));
   }
 
   @JsonKey(ignore: true)
@@ -181,7 +201,8 @@ class _$_TweetMeta implements _TweetMeta {
       const DeepCollectionEquality().hash(newestId),
       const DeepCollectionEquality().hash(oldestId),
       const DeepCollectionEquality().hash(resultCount),
-      const DeepCollectionEquality().hash(nextToken));
+      const DeepCollectionEquality().hash(nextToken),
+      const DeepCollectionEquality().hash(previousToken));
 
   @JsonKey(ignore: true)
   @override
@@ -196,10 +217,12 @@ class _$_TweetMeta implements _TweetMeta {
 
 abstract class _TweetMeta implements TweetMeta {
   const factory _TweetMeta(
-      {@JsonKey(name: 'newest_id') final String? newestId,
-      @JsonKey(name: 'oldest_id') final String? oldestId,
-      @JsonKey(name: 'result_count') final int? resultCount,
-      @JsonKey(name: 'next_token') final String? nextToken}) = _$_TweetMeta;
+          {@JsonKey(name: 'newest_id') final String? newestId,
+          @JsonKey(name: 'oldest_id') final String? oldestId,
+          @JsonKey(name: 'result_count') final int? resultCount,
+          @JsonKey(name: 'next_token') final String? nextToken,
+          @JsonKey(name: 'previous_token') final String? previousToken}) =
+      _$_TweetMeta;
 
   factory _TweetMeta.fromJson(Map<String, dynamic> json) =
       _$_TweetMeta.fromJson;
@@ -216,6 +239,9 @@ abstract class _TweetMeta implements TweetMeta {
   @override
   @JsonKey(name: 'next_token')
   String? get nextToken => throw _privateConstructorUsedError;
+  @override
+  @JsonKey(name: 'previous_token')
+  String? get previousToken => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
   _$TweetMetaCopyWith<_TweetMeta> get copyWith =>