+++ lib/src/service/users/user_meta.dart
@@ -11,8 +11,9 @@ part 'user_meta.g.dart';
 @freezed
 class UserMeta with _$UserMeta {
   const factory UserMeta({
-    @JsonKey(name: 'result_count') required int resultCount,
-    @JsonKey(name: 'next_token') required String nextToken,
+    @JsonKey(name: 'result_count') int? resultCount,
+    @JsonKey(name: 'next_token') String? nextToken,
+    @JsonKey(name: 'previous_token') String? previousToken,
   }) = _UserMeta;
 
   factory UserMeta.fromJson(Map<String, Object?> json) =>