+++ lib/src/service/users/user_meta.freezed.dart
@@ -21,9 +21,11 @@ UserMeta _$UserMetaFromJson(Map<String, dynamic> json) {
 /// @nodoc
 mixin _$UserMeta {
   @JsonKey(name: 'result_count')
-  int get resultCount => throw _privateConstructorUsedError;
+  int? get resultCount => throw _privateConstructorUsedError;
   @JsonKey(name: 'next_token')
-  String get nextToken => throw _privateConstructorUsedError;
+  String? get nextToken => throw _privateConstructorUsedError;
+  @JsonKey(name: 'previous_token')
+  String? get previousToken => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
@@ -36,8 +38,9 @@ abstract class $UserMetaCopyWith<$Res> {
   factory $UserMetaCopyWith(UserMeta value, $Res Function(UserMeta) then) =
       _$UserMetaCopyWithImpl<$Res>;
   $Res call(
-      {@JsonKey(name: 'result_count') int resultCount,
-      @JsonKey(name: 'next_token') String nextToken});
+      {@JsonKey(name: 'result_count') int? resultCount,
+      @JsonKey(name: 'next_token') String? nextToken,
+      @JsonKey(name: 'previous_token') String? previousToken});
 }
 
 /// @nodoc
@@ -52,16 +55,21 @@ class _$UserMetaCopyWithImpl<$Res> implements $UserMetaCopyWith<$Res> {
   $Res call({
     Object? resultCount = freezed,
     Object? nextToken = freezed,
+    Object? previousToken = freezed,
   }) {
     return _then(_value.copyWith(
       resultCount: resultCount == freezed
           ? _value.resultCount
           : resultCount // ignore: cast_nullable_to_non_nullable
-              as int,
+              as int?,
       nextToken: nextToken == freezed
           ? _value.nextToken
           : nextToken // ignore: cast_nullable_to_non_nullable
-              as String,
+              as String?,
+      previousToken: previousToken == freezed
+          ? _value.previousToken
+          : previousToken // ignore: cast_nullable_to_non_nullable
+              as String?,
     ));
   }
 }
@@ -72,8 +80,9 @@ abstract class _$UserMetaCopyWith<$Res> implements $UserMetaCopyWith<$Res> {
       __$UserMetaCopyWithImpl<$Res>;
   @override
   $Res call(
-      {@JsonKey(name: 'result_count') int resultCount,
-      @JsonKey(name: 'next_token') String nextToken});
+      {@JsonKey(name: 'result_count') int? resultCount,
+      @JsonKey(name: 'next_token') String? nextToken,
+      @JsonKey(name: 'previous_token') String? previousToken});
 }
 
 /// @nodoc
@@ -89,16 +98,21 @@ class __$UserMetaCopyWithImpl<$Res> extends _$UserMetaCopyWithImpl<$Res>
   $Res call({
     Object? resultCount = freezed,
     Object? nextToken = freezed,
+    Object? previousToken = freezed,
   }) {
     return _then(_UserMeta(
       resultCount: resultCount == freezed
           ? _value.resultCount
           : resultCount // ignore: cast_nullable_to_non_nullable
-              as int,
+              as int?,
       nextToken: nextToken == freezed
           ? _value.nextToken
           : nextToken // ignore: cast_nullable_to_non_nullable
-              as String,
+              as String?,
+      previousToken: previousToken == freezed
+          ? _value.previousToken
+          : previousToken // ignore: cast_nullable_to_non_nullable
+              as String?,
     ));
   }
 }
@@ -107,22 +121,26 @@ class __$UserMetaCopyWithImpl<$Res> extends _$UserMetaCopyWithImpl<$Res>
 @JsonSerializable()
 class _$_UserMeta implements _UserMeta {
   const _$_UserMeta(
-      {@JsonKey(name: 'result_count') required this.resultCount,
-      @JsonKey(name: 'next_token') required this.nextToken});
+      {@JsonKey(name: 'result_count') this.resultCount,
+      @JsonKey(name: 'next_token') this.nextToken,
+      @JsonKey(name: 'previous_token') this.previousToken});
 
   factory _$_UserMeta.fromJson(Map<String, dynamic> json) =>
       _$$_UserMetaFromJson(json);
 
   @override
   @JsonKey(name: 'result_count')
-  final int resultCount;
+  final int? resultCount;
   @override
   @JsonKey(name: 'next_token')
-  final String nextToken;
+  final String? nextToken;
+  @override
+  @JsonKey(name: 'previous_token')
+  final String? previousToken;
 
   @override
   String toString() {
-    return 'UserMeta(resultCount: $resultCount, nextToken: $nextToken)';
+    return 'UserMeta(resultCount: $resultCount, nextToken: $nextToken, previousToken: $previousToken)';
   }
 
   @override
@@ -132,7 +150,9 @@ class _$_UserMeta implements _UserMeta {
             other is _UserMeta &&
             const DeepCollectionEquality()
                 .equals(other.resultCount, resultCount) &&
-            const DeepCollectionEquality().equals(other.nextToken, nextToken));
+            const DeepCollectionEquality().equals(other.nextToken, nextToken) &&
+            const DeepCollectionEquality()
+                .equals(other.previousToken, previousToken));
   }
 
   @JsonKey(ignore: true)
@@ -140,7 +160,8 @@ class _$_UserMeta implements _UserMeta {
   int get hashCode => Object.hash(
       runtimeType,
       const DeepCollectionEquality().hash(resultCount),
-      const DeepCollectionEquality().hash(nextToken));
+      const DeepCollectionEquality().hash(nextToken),
+      const DeepCollectionEquality().hash(previousToken));
 
   @JsonKey(ignore: true)
   @override
@@ -155,18 +176,22 @@ class _$_UserMeta implements _UserMeta {
 
 abstract class _UserMeta implements UserMeta {
   const factory _UserMeta(
-          {@JsonKey(name: 'result_count') required final int resultCount,
-          @JsonKey(name: 'next_token') required final String nextToken}) =
+          {@JsonKey(name: 'result_count') final int? resultCount,
+          @JsonKey(name: 'next_token') final String? nextToken,
+          @JsonKey(name: 'previous_token') final String? previousToken}) =
       _$_UserMeta;
 
   factory _UserMeta.fromJson(Map<String, dynamic> json) = _$_UserMeta.fromJson;
 
   @override
   @JsonKey(name: 'result_count')
-  int get resultCount => throw _privateConstructorUsedError;
+  int? get resultCount => throw _privateConstructorUsedError;
   @override
   @JsonKey(name: 'next_token')
-  String get nextToken => throw _privateConstructorUsedError;
+  String? get nextToken => throw _privateConstructorUsedError;
+  @override
+  @JsonKey(name: 'previous_token')
+  String? get previousToken => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
   _$UserMetaCopyWith<_UserMeta> get copyWith =>